Validaciones, Strings y permisos de tablas de GG20031
Se aplico los permisos de usuario a las tablas de Misael [GG20031]. Asimismo se ingreso los string necesario para los mensajes de los meotdos respectivos de las tablas asignadas a GG20031. También se hizo las validaciones pertinentes en dichos layouts y actividades.

+++ app/src/main/java/com/example/g10proyecto01/ControlG10Proyecto01.java
@@ -205,9 +205,10 @@ public class ControlG10Proyecto01 {
     }
     /******************************************** Tabla EmpleadoUES ********************************************/
     // CAMPOS: {"id_empleado", "id_tipo_empleado", "nombre_empleado", "apellido_empleado", "email_empleado", "telefono_empleado"}
+
     /*  Insertar EmpleadoUES  */
     public String insertar(EmpleadoUES empleadoUES){
-        String regInsertados="Registro Insertado Nº= ";
+        String regInsertados = context.getResources().getString(R.string.regInsertado);
         long contador = 0;
         if (verificarIntegridad(empleadoUES, 1)) {
             ContentValues empleado = new ContentValues();
@@ -221,7 +222,7 @@ public class ControlG10Proyecto01 {
         }
         if(contador == -1 || contador == 0)
         {
-            regInsertados= "Error al Insertar el registro, Registro Duplicado. Verificar inserción";
+            regInsertados= context.getResources().getString(R.string.error);
         }
         else {
             regInsertados=regInsertados+contador;
@@ -258,20 +259,22 @@ public class ControlG10Proyecto01 {
         empleado.put("email_empleado", empleadoUES.getEmail_empleado());
         empleado.put("telefono_empleado", empleadoUES.getTelefono_empleado());
         db.update("Empleado_UES", empleado, "id_empleado = ?", id);
-        return "Registro Actualizado Correctamente";
+        return context.getResources().getString(R.string.regActualizado);
     }
 
-    /*  Inserta EmpleadoUES  */
+    /*  Eliminar EmpleadoUES  */
     public String eliminar(EmpleadoUES empleadoUES){
-        String regAfectados="filas afectadas= ";
+        String regAfectados=context.getResources().getString(R.string.regEliminados);
         int contador=0;
-        //if (verificarIntegridad(tipoEmpleado, 2)){
+        if (verificarIntegridad(empleadoUES, 5)){
             String where="id_empleado=" + empleadoUES.getId_empleado();
             contador += db.delete("Empleado_UES", where, null);
             regAfectados+=contador;
             return regAfectados;
-        //}
-        //else return "Hay una referencia en Empleado UES de id tipo de empleado " + tipoEmpleado.getId_tipo_empleado();
+        }
+        else return context.getResources().getString(R.string.referencia) + " " +
+                    context.getResources().getString(R.string.IdEmpleado) + " " +
+                    empleadoUES.getId_empleado();
     }
 
     /******************************************** Tabla Docente ********************************************/
@@ -279,7 +282,7 @@ public class ControlG10Proyecto01 {
 
     /*  Insertar Docente  */
     public String insertar(Docente docente){
-        String regInsertados="Registro Insertado Nº= ";
+        String regInsertados=context.getResources().getString(R.string.regInsertado);
         long contador=0;
         if (verificarIntegridad(docente, 3)){
             ContentValues docent = new ContentValues();
@@ -291,7 +294,7 @@ public class ControlG10Proyecto01 {
         }
         if(contador==-1 || contador==0)
         {
-            regInsertados= "Error al Insertar el registro, Registro Duplicado. Verificar inserción";
+            regInsertados= context.getResources().getString(R.string.error);
         }
         else {
             regInsertados=regInsertados+contador;
@@ -323,21 +326,23 @@ public class ControlG10Proyecto01 {
         docent.put("nip_docente", docente.getNip_docente());
         docent.put("categoria_docente", docente.getCategoria_docente());
         db.update("Docente",docent,"id_docente= ?", id);
-        return "Registro Actualizado Correctamente";
+        return context.getResources().getString(R.string.regActualizado);
     }
 
 
     /*  Eliminar Docente  */
     public String eliminar(Docente docente){
-        String regAfectados="filas afectadas= ";
+        String regAfectados=context.getResources().getString(R.string.regEliminados);
         int contador=0;
-        //if (verificarIntegridad(docente, 4)) {
+        if (verificarIntegridad(docente, 4)) {
             String where = "id_Docente=" + docente.getId_docente();
             contador += db.delete("Docente", where, null);
             regAfectados += contador;
             return regAfectados;
-        //}
-        //else return "Hay una referencia en Empleado UES de id tipo de empleado " + docente.getId_docente();
+        }
+        else return context.getResources().getString(R.string.referencia) + " " +
+                    context.getResources().getString(R.string.IdDocente) + " " +
+                    docente.getId_docente();
     }
 
     /******************************************** Tabla TipoEmpleado ********************************************/
@@ -345,17 +350,16 @@ public class ControlG10Proyecto01 {
 
     /*  Insertar EmpleadoUES  */
     public String insertar(TipoEmpleado tipoEmpleado){
-        String regInsertados="Registro Insertado Nº= ";
+        String regInsertados=context.getResources().getString(R.string.regInsertado);
         long contador=0;
         ContentValues tEmpleado = new ContentValues();
         tEmpleado.put("id_tipo_empleado", tipoEmpleado.getId_tipo_empleado());
         tEmpleado.put("ocupacion", tipoEmpleado.getOcupacion());
 
-        ;
         contador = db.insert("Tipo_de_Empleado", null, tEmpleado);
         if(contador==-1 || contador==0)
         {
-            regInsertados= "Error al Insertar el registro, Registro Duplicado. Verificar inserción";
+            regInsertados= context.getResources().getString(R.string.error);
         }
         else {
             regInsertados=regInsertados+contador;
@@ -383,12 +387,12 @@ public class ControlG10Proyecto01 {
         ContentValues cv = new ContentValues();
         cv.put("ocupacion", tipoEmpleado.getOcupacion());
         db.update("Tipo_de_Empleado",cv,"id_tipo_empleado = ?", id);
-        return "Registro Actualizado Correctamente";
+        return context.getResources().getString(R.string.regActualizado);
     }
 
     /*  Eliminar tipoEmpleado  */
     public String eliminar(TipoEmpleado tipoEmpleado){
-        String regAfectados="filas afectadas= ";
+        String regAfectados=context.getResources().getString(R.string.regEliminados);
         int contador=0;
         if (verificarIntegridad(tipoEmpleado, 2)){
             String where="id_tipo_empleado=" + tipoEmpleado.getId_tipo_empleado();
@@ -396,7 +400,9 @@ public class ControlG10Proyecto01 {
             regAfectados+=contador;
             return regAfectados;
         }
-        else return "Hay una referencia en Empleado UES de id tipo de empleado " + tipoEmpleado.getId_tipo_empleado();
+        else return context.getResources().getString(R.string.referencia) + " " +
+                    context.getResources().getString(R.string.tvIdTipoEmpleado) + " " +
+                    tipoEmpleado.getId_tipo_empleado();
     }
     /******************************************** Tabla OpcionCrud ********************************************/
     // CAMPOS: {"id_opcion_crud", "des_opcion"}
@@ -649,7 +655,7 @@ public class ControlG10Proyecto01 {
                 return false;
             }
             case 2: {
-                //VERIFICA QUE NO SE ELIMINE UN TIPO EMPLEADO MIENTRA EXiISTA UNA REFERENCIA DE EL EN LA TABLA EMPLEADO UES
+                //VERIFICA QUE NO SE ELIMINE UN TIPO EMPLEADO MIENTRA EXISTA UNA REFERENCIA DE EL EN LA TABLA EMPLEADO UES
                 TipoEmpleado tipoEmpleado = (TipoEmpleado) dato;
                 Cursor cursor = db.query(true, "Empleado_UES", new String[]{"id_tipo_empleado"}, "id_tipo_empleado ='" + tipoEmpleado.getId_tipo_empleado() + "'", null, null, null, null, null);
                 if (cursor.moveToFirst())
@@ -670,6 +676,26 @@ public class ControlG10Proyecto01 {
                 }
                 return false;
             }
+            case 4: {
+                //VERIFICA QUE NO SE ELIMINE UN DOCENTE MIENTRA EXISTA UNA REFERENCIA DE EL EN LA TABLA OFERTA ACADEMICA
+                Docente docente = (Docente) dato;
+                Cursor cursor = db.query(true, "Oferta_Academica", new String[]{"id_docente"}, "id_docente ='" + docente.getId_docente() + "'", null, null, null, null, null);
+                if (cursor.moveToFirst())
+                    return false;
+                else
+                    return true;
+            }
+            case 5:{
+                //VERIFICA QUE NO SE ELIMINE UN EMPLEADO UES MIENTRA EXISTA UNA REFERENCIA DE EL EN LAS TABLAS DOCENTE Y LOCAL ADMINISTRADO
+                EmpleadoUES empleado = (EmpleadoUES) dato;
+                Cursor cursor1 = db.query(true, "Docente", new String[]{"id_empleado"}, "id_empleado ='" + empleado.getId_empleado() + "'", null, null, null, null, null);
+                Cursor cursor2 = db.query(true, "Local_Administrado", new String[]{"id_empleado"}, "id_empleado ='" + empleado.getId_empleado() + "'", null, null, null, null, null);
+                if (cursor1.moveToFirst() || cursor2.moveToFirst())
+                    return false;
+                else
+                    return true;
+
+            }
             default: return false;
         }
     }
@@ -680,13 +706,14 @@ public class ControlG10Proyecto01 {
 
     //Usuarios iniciales
     public void permisosUsuarios(){
+        db.execSQL("DELETE FROM Usuario");
+        db.execSQL("DELETE FROM OpcionCrud");
+        db.execSQL("DELETE FROM AccesoUsuario");
 
+        //Usuarios
         final String[] IDusuario = {"01", "02", "03", "04", "05"};
         final String[] nomUsuario = {"Misael", "Fabio", "Claudia", "Leonardo", "Alexander"};
         final String[] clave = {"GG20031", "FM19038", "AC17033", "EL19004", "HS19011"};
-
-
-        db.execSQL("DELETE FROM Usuario");
 
         Usuario user = new Usuario();
 
@@ -696,6 +723,27 @@ public class ControlG10Proyecto01 {
             user.setClave(clave[i]);
             insertar(user);
         }
+
+        //ACCESO USUARIO
+
+        final int[] idsAccesoUsuario= {1,2,3,4,5,6,7,8};
+        final String[] IDusuarios = {"01", "02", "03", "04","05", "01", "01", "01"};
+        final int [] idOpcionCrud_Access = {1,2,3,4,1,2,3,4};
+        for (int i = 0; i < idsAccesoUsuario.length; i++) {
+            AccesoUsuario accesoUsuario = new AccesoUsuario(idsAccesoUsuario[i],IDusuarios[i],idOpcionCrud_Access[i]);
+            insertar(accesoUsuario);
+        }
+
+        //OPCIONCRUD
+
+        final int[] idOpcionCrud= {1,2,3,4};
+        final String[] opcionCrud = {"Insertar","Actualizar","Eliminar","Consultar"};
+        for (int i = 0; i < opcionCrud.length; i++) {
+            OpcionCrud opcion = new OpcionCrud(idOpcionCrud[i],opcionCrud[i]);
+            insertar(opcion);
+        }
+
+
     }
 
     // Método para rescatar datos de de la base para el Spinner
@@ -723,8 +771,6 @@ public class ControlG10Proyecto01 {
         db.execSQL("DELETE FROM Evento_Especial");
 
         //CICLO
-        //DOCENTE
-        //EMPLEADO_UES
         //ESCUELA
         final int[] EscuelaId = {1,2,3,4};
         final String[] EscuelaAcronimo = {"EISI","UCB","EII","EIE"};
@@ -760,17 +806,25 @@ public class ControlG10Proyecto01 {
         //MATERIA
 
         //OFERTAACADEMICA
+        //PROPUESTAESPECIFICA
+        //PROPUESTAGENERAL
+
+        //ACCESO USUARIO
+        final int[] idsAccesoUsuario= {1,2,3,4,5,6,7,8};
+        final String[] IDusuarios = {"01", "02", "03", "04","05", "01", "01", "01"};
+        final int [] idOpcionCrud_Access = {1,2,3,4,1,2,3,4};
+        for (int i = 0; i < idsAccesoUsuario.length; i++) {
+            AccesoUsuario accesoUsuario = new AccesoUsuario(idsAccesoUsuario[i],IDusuarios[i],idOpcionCrud_Access[i]);
+            insertar(accesoUsuario);
+        }
 
         //OPCIONCRUD
         final int[] idOpcionCrud= {1,2,3,4};
-        final String[] opcionCrud = {"insertar","actualizar","eliminar","ver detalle"};
+        final String[] opcionCrud = {"Insertar","Actualizar","Eliminar","Consultar"};
         for (int i = 0; i < opcionCrud.length; i++) {
             OpcionCrud opcion = new OpcionCrud(idOpcionCrud[i],opcionCrud[i]);
             insertar(opcion);
         }
-
-        //PROPUESTAESPECIFICA
-        //PROPUESTAGENERAL
 
         //TIPO EMPLEADO
         final int[] idTipoEmpleado = {1,2,3,4};
@@ -817,14 +871,6 @@ public class ControlG10Proyecto01 {
             insertar(docente);
         }
 
-        //ACCESO USUARIO
-        final int[] idsAccesoUsuario= {1,2,3,4};
-        final String[] IDusuarios = {"01", "02", "03", "04","05"};
-        for (int i = 0; i < idsAccesoUsuario.length; i++) {
-            AccesoUsuario accesoUsuario = new AccesoUsuario(idsAccesoUsuario[i],IDusuarios[i],idsAccesoUsuario[i]);
-            insertar(accesoUsuario);
-        }
-
         //TIPO DE EVENTO
         final int[] idTipoEvento = {1,2,3,4};
         final String[] nombre = {"Foro", "Conferencia", "Examen Parcial", "Capacitacion"};
@@ -838,7 +884,8 @@ public class ControlG10Proyecto01 {
 
         cerrar();
 
-        return "Guardo Correctamente";
+        return context.getResources().getString(R.string.llenadoBD);
+        //return "LLenado bien";
     }
 }
